requirements.txt
accelerate==0.29.3
bitsandbytes==0.42.0
brotli==1.0.9
chardet==5.2.0
datasets==2.19.0
docx2txt==0.8
einops==0.8.0
greenlet==3.0.3
jupyterlab==4.0.11
langchain==0.1.16
llama-index==0.10.32
llama-index-embeddings-langchain==0.1.2
llama-index-llms-huggingface
pip-chill==1.0.3
pysocks==1.7.1
pytelegrambotapi==4.18.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
torchaudio==2.2.0.dev
torchvision
